# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/internal/shopify/app-uninstalled": {
      "post": {
        "operationId": "InternalShopifyController_appUninstalled",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopifyShopEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Handle verified Shopify app uninstall",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/v1/internal/shopify/shop-redact": {
      "post": {
        "operationId": "InternalShopifyController_shopRedact",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopifyShopEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Permanently redact all ACA Locale shop data",
        "tags": [
          "Internal"
        ]
      }
    },
...
  },
  "info": {
    "title": "ACA Locale API",
    "description": "Technical API documentation for ACA Locale.\n\nACA Locale is an AI-powered localization management platform.\n\nThis specification is generated automatically from the NestJS backend.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ShopifyShopEventDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}